C99isms for printf format strings.

+++ src/libfsm/print/go.c
@@ -9,6 +9,7 @@
 #include <stdint.h>
 #include <errno.h>
 #include <stdio.h>
+#include <inttypes.h>
 
 #include <print/esc.h>
 
@@ -64,7 +65,7 @@ cmp_operator(int cmp)
 static void
 print_label(FILE *f, const struct dfavm_op_ir *op, const struct fsm_options *opt)
 {
-	fprintf(f, "l%lu:", (unsigned long) op->index);
+	fprintf(f, "l%" PRIu32 ":", op->index);
 
 	if (op->ir_state->example != NULL) {
 		fprintf(f, " // e.g. \"");
@@ -99,14 +100,14 @@ print_end(FILE *f, const struct dfavm_op_ir *op, const struct fsm_options *opt,
 	if (opt->endleaf != NULL) {
 		opt->endleaf(f, op->ir_state->end_ids, opt->endleaf_opaque);
 	} else {
-		fprintf(f, "{\n\t\treturn %lu\n\t}\n", (unsigned long) (op->ir_state - ir->states));
+		fprintf(f, "{\n\t\treturn %td\n\t}\n", op->ir_state - ir->states);
 	}
 }
 
 static void
 print_branch(FILE *f, const struct dfavm_op_ir *op)
 {
-	fprintf(f, "{\n\t\tgoto l%lu\n\t}\n", (unsigned long) op->u.br.dest_arg->index);
+	fprintf(f, "{\n\t\tgoto l%" PRIu32 "\n\t}\n", op->u.br.dest_arg->index);
 }
 
 static void
